Rename package `js` to `javascript`
