+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/javascript/plugins/Plugins.kt
@@ -24,10 +24,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.internal.gradle.js.plugins
+package io.spine.internal.gradle.javascript.plugins
 
-import io.spine.internal.gradle.js.JsContext
-import io.spine.internal.gradle.js.JsEnvironment
+import io.spine.internal.gradle.javascript.JsContext
+import io.spine.internal.gradle.javascript.JsEnvironment
 import org.gradle.api.Project
 import org.gradle.api.tasks.TaskContainer
